FIX: Product page - Move js from main app.js file to product page file.
